Operators Chat Group FE v0 (#90)
I've added the FE to our codebase in `webapp/main`. If you all have any
ideas for a better name feel free to suggest.

In any case, the application can be spun up locally via
```make local-webapp-main```

For local development, this shares the same sqlite database as the concrete API, so I would recommend spinning that up to manipulate the database via
```make local-api```

You'll need the refresh the FE to see any changes. Currently there's
only dynamic loading for orchestrators since messages are hardcoded and
the API for it doesn't exist yet.


![image](https://github.com/user-attachments/assets/76d23cb5-d667-4314-8acf-c945f10e83f7)

![image](https://github.com/user-attachments/assets/99d05ac4-c805-4921-9d4c-9511e89ddde2)