Marketing pages for concrete and agentserver (#209)
First iteration on landing page sections for concrete and agentserver. Improved responsiveness of navbar

Focus on the content, not on the presentation for now; the main files
are `concrete.html`, `agentserver.html`, and `about.html`.
Feel free to leave critiques and suggestions.

I'm also pondering a more interesting name than agentserver. It's
descriptive, but not that slick.
Ideas:
- StAItion
- Operatus (Top Contender)
- OpsPlex
- Orchestrator
- Could reuse Tress here
- OperatorCore
- OperatorHub
- TOP (The Operations Platform)

---------

Co-authored-by: kgang <kent.y.gang@gmail.com>